requirements.txt
nltk==3.8.1
Pillow==10.2.0 